Unknown: override _Clone()
Add an override for the _Clone() method to make it also increase the COM reference count.
This ensures that, if a clone is created, releasing the original or the clone does NOT make the other one stop working.
The discussion on this can be found at http://www.autohotkey.com/forum/viewtopic.php?p=517524#517524.
This *might* also solve #46.